# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/beneficiary-payouts": {
      "post": {
        "description": "Initie un paiement vers un bénéficiaire (prestataire, fournisseur, salarié, etc.). Réponse immédiate en statut en attente ; le traitement est asynchrone côté fournisseur (Wave, SPI). Les mises à jour arrivent via webhooks. Sert à payer des tiers depuis le solde du compte.",
        "operationId": "BeneficiaryPayoutsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Paiement initié avec succès (statut en attente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficiaryPayoutResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou solde insuffisant"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lancer un paiement vers un bénéficiaire",
        "tags": ["Paiements vers bénéficiaires"]
      },
      "get": {
        "description": "Renvoie les paiements de l'organisation avec pagination et filtres optionnels (statuts, plage de dates, devise).",
        "operationId": "BeneficiaryPayoutsController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Décalage pour la pagination",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximal de résultats",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "currencyCode",
            "required": false,
            "in": "query",
            "description": "Code devise",
            "schema": {
              "example": "XOF",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Date de fin (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Date de début (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statuses",
            "required": false,
            "in": "query",
            "description": "Statuts (séparés par des virgules)",
            "schema": {
              "example": "pending,completed,failed",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des paiements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BeneficiaryPayoutResponseDto"
                      }
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les paiements vers bénéficiaires",
        "tags": ["Paiements vers bénéficiaires"]
      }
    },
    "/beneficiary-payouts/{id}": {
      "get": {
        "description": "Renvoie le détail d’un paiement (statut actuel et informations fournisseur).",
        "operationId": "BeneficiaryPayoutsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du paiement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du paiement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficiaryPayoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Paiement introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un paiement bénéficiaire par ID",
        "tags": ["Paiements vers bénéficiaires"]
      }
    },
...
    "/charge/wave": {
      "post": {
        "operationId": "ChargesController_createWaveCharge",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Encaissement initié (corps JSON renvoyé par la fonction edge Wave)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou erreur API Wave"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lancer un encaissement direct Wave",
        "tags": ["Encaissements"]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "API de traitement des paiements pour les entreprises d'Afrique de l'Ouest francophone.",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "BeneficiaryPayoutResponseDto": {
        "type": "object",
        "properties": {
          "account_id": {
            "type": "string",
            "example": "string"
          },
          "amount": {
            "type": "number",
            "example": 123
          },
          "created_at": {
            "type": "string",
            "example": "string"
          },
          "created_by": {
            "type": "string",
            "example": "string"
          },
          "currency_code": {
            "type": "string",
            "example": "string"
          },
          "metadata": {
            "type": "object",
            "example": {}
          },
          "organization_id": {
            "type": "string",
            "example": "string"
          },
          "payout_id": {
            "type": "string",
            "example": "string"
          },
          "payout_method_id": {
            "type": "string",
            "example": "string"
          },
          "provider_code": {
            "type": "string",
            "example": "string"
          },
          "spi_bulk_instruction_id": {
            "type": "string",
            "example": "string"
          },
          "status": {
            "type": "string",
            "example": "string"
          },
          "updated_at": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "account_id",
          "amount",
          "created_at",
          "created_by",
          "currency_code",
          "metadata",
          "organization_id",
          "payout_id",
          "payout_method_id",
          "provider_code",
          "spi_bulk_instruction_id",
          "status",
          "updated_at"
        ]
      },
...
    }
  }
}